refactor: Seperate Advisor and Client classes, refactor code
