+++ src/main/java/com/wellsfargo/counselor/entity/Advisor.java
@@ -85,86 +85,3 @@ public class Advisor {
 
 }
 
-@Entity
-public class Client {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    @ManyToOne
-    @JoinColumn(name = "advisor_id", nullable = false)
-    private Advisor advisor;
-
-    @OneToOne(mappedBy = "client", cascade = CascadeType.ALL)
-    private Portfolio portfolio;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(name = "contact_info")
-    private String contactInfo;
-
-    @Column(name = "created_at")
-    private LocalDateTime createdAt;
-
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
-
-    public Client(Advisor advisor, Portfolio portfolio, String name, String contactInfo,
-                  LocalDateTime createdAt, LocalDateTime updatedAt) {
-        this.advisor = advisor;
-        this.portfolio = portfolio;
-        this.name = name;
-        this.contactInfo = contactInfo;
-        this.createdAt = createdAt;
-        this.updatedAt = updatedAt;
-    }
-
-    protected Client() {
-    }
-
-    public Long getId(){
-        return id;
-    }
-
-    public Advisor getAdvisor() {
-        return advisor;
-    }
-
-    public void setAdvisor(Advisor advisor) {
-        this.advisor = advisor;
-    }
-
-    public Portfolio getPortfolio() {
-        return portfolio;
-    }
-
-    public String getName(){
-        return name;
-    }
-
-    public String getContactInfo() {
-        return contactInfo;
-    }
-
-    public void setContactInfo(String contactInfo) {
-        this.contactInfo = contactInfo;
-    }
-
-    public LocalDateTime getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(LocalDateTime createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public LocalDateTime getUpdatedAt() {
-        return updatedAt;
-    }
-
-    public void setUpdatedAt(LocalDateTime updatedAt) {
-        this.updatedAt = updatedAt;
-    }
-}
